solr9cores/dataimport/requirements.txt
Flask==3.0.0
JayDeBeApi==1.2.3
beautifulsoup4==4.12.2
lxml==4.9.3
pysolr==3.9.0
boto3==1.28.62
python-decouple==3.8